% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lib.R
\name{plot.straineff.ci}
\alias{plot.straineff.ci}
\title{Plot approximate posterior credible intervals of additive haplotype effects as a caterpillar plot}
\usage{
\method{plot}{straineff.ci}(inla.diploffect.ci, sn = NULL,
  xlab = "Haplotype Effects", main = NULL, main.cex = 1,
  main.line = 2, include.top.axis = TRUE, flip = TRUE, ...)
}
\arguments{
\item{inla.diploffect.ci}{A Diploffect.INLA summary object. Output from run.diploffect.inla.summary.stats(). Contains credible information
for various model parameters.}

\item{sn}{DEFAULT: NULL. Strain names to be used as labels in the plot. Will default to the names used in the Diploffect.INLA model probabilities.}

\item{xlab}{DEFAULT: "Haplotype Effects". Label for the x axis of the plot.}

\item{main}{DEFAULT: "". Title for the plot.}

\item{flip}{DEFAULT: TRUE. Flips the order of the haplotypes. Allows a little flexibility in how the effects are arranged.}
}
\value{
Nothing. Produces plot.
}
\description{
This function takes a Diploffect.INLA summary object and plots the posterior credible intervals of
the additive haplotype effects.
}
\examples{
library(Diploffect.INLA)
data(exampleCC)
data(locusmatrix)
inla.diploffect <- run.diploffect.inla(formula=y~1+(1|strain)+(1|dose.date), add.on=FALSE, data=exampleCC, K=NULL,
                                       prob.matrix=locusmatrix,
                                       num.draws=10, 
                                       use.dip.lincomb=TRUE, seed=1, gamma.rate=1, impute.on="CCline")
inla.diploffect.summary <- run.diploffect.inla.summary.stats(inla.diploffect)
plot.straineff.ci(inla.diploffect.summary, flip=FALSE)
}
