% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lib.R
\name{plot.varexp.ci}
\alias{plot.varexp.ci}
\title{Plot approximate posterior credible intervals of proportion variance explained (PVE) as a caterpillar plot}
\usage{
\method{plot}{varexp.ci}(inla.diploffect.ci, xlab = "Variance Explained",
  main = NULL, main.cex = 1, main.line = 2, add.numbers = FALSE,
  ...)
}
\arguments{
\item{inla.diploffect.ci}{A Diploffect.INLA summary object. Output from run.diploffect.inla.summary.stats(). Contains credible information
for various model parameters.}

\item{xlab}{DEFAULT: "Variance Explained". Label for the x axis of the plot.}

\item{add.numbers}{DEFAULT: FALSE. Adds numeric summaries next to intervals.}
}
\value{
Nothing. Produces plot.
}
\description{
This function takes a Diploffect.INLA summary object and plots the posterior credible intervals of
the proportion of the variance explained.
}
\examples{
library(Diploffect.INLA)
data(exampleCC)
data(locusmatrix)
inla.diploffect <- run.diploffect.inla(formula=y~1+(1|strain)+(1|dose.date), add.on=FALSE, data=exampleCC, K=NULL,
                                       prob.matrix=locusmatrix,
                                       num.draws=10, 
                                       use.dip.lincomb=TRUE, seed=1, gamma.rate=1, impute.on="CCline")
inla.diploffect.summary <- run.diploffect.inla.summary.stats(inla.diploffect)
plot.varexp.ci(inla.diploffect.summary, add.numbers=TRUE)
}
