% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lib.R
\name{plot_qtl.varexp.ci}
\alias{plot_qtl.varexp.ci}
\title{Plot approximate posterior credible intervals of proportion variance explained (PVE) due to QTL as a caterpillar plot}
\usage{
plot_qtl.varexp.ci(inla.diploffect.ci,
  xlab = "QTL Effects (Variance Explained)", ...)
}
\arguments{
\item{inla.diploffect.ci}{A Diploffect.INLA summary object. Output from run.diploffect.inla.summary.stats(). Contains credible information
for various model parameters.}

\item{xlab}{DEFAULT: "QTL Effects (Variance Explained)". Label for the x axis of the plot.}
}
\value{
Nothing. Produces plot.
}
\description{
This function takes a Diploffect.INLA summary object and plots the posterior credible intervals of
the proportion of the variance explained due to QTL effects.
}
\examples{
library(Diploffect.INLA)
data(exampleCC)
data(locusmatrix)
inla.diploffect <- run.diploffect.inla(formula=y~1+(1|strain)+(1|dose.date), add.on=FALSE, data=exampleCC, K=NULL,
                                       prob.matrix=locusmatrix,
                                       num.draws=10, 
                                       use.dip.lincomb=TRUE, seed=1, gamma.rate=1, impute.on="CCline")
inla.diploffect.summary <- run.diploffect.inla.summary.stats(inla.diploffect)
plot_qtl.varexp.ci(inla.diploffect.summary)
}
