% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.diploffect.inla.genomecache.R
\name{run.diploffect.inla.through.genomecache}
\alias{run.diploffect.inla.through.genomecache}
\title{Run Diploffect model on diplotype (haplotype pair) probabilities stored in a genomecache directory}
\usage{
run.diploffect.inla.through.genomecache(formula, data, K = NULL, genomecache,
  locus, num.draws, use.dip.lincomb = TRUE, seed = 1, gamma.rate = 1,
  impute.on = "SUBJECT.NAME", weights.on = NULL, scale.response = TRUE,
  do.augment.of.cache = FALSE, return.joint.posterior.samples = FALSE)
}
\arguments{
\item{formula}{An lm style formula with functions of outcome and covariates in data frame.}

\item{data}{A data frame with outcome and potential covariates. Should also have individual IDs
that link to IDs in the genomecache, commonly with this column named "SUBJECT.NAME".}

\item{K}{A positive semi-definite relationship matrix, usually a realized genetic relationship matrix (GRM)
based on SNP genotypes are the haplotype probabilistic reconstructions. Colnames and rownames should match
the SUBJECT.NAME column in the data frame.}

\item{genomecache}{The path to the genomecache directory. The genomecache is a particularly structured
directory that stores the haplotype probabilities/dosages at each locus. It has an additive model
subdirectory and a full model subdirectory. Each contains subdirectories for each chromosome, which then
store .RData files for the probabilities/dosages.}

\item{locus}{The locus for which to run Diploffect. It must be contained in the genomecache.}

\item{num.draws}{The number of times the diplotype (haplotype pair) probabilities are sampled, and
INLA is run.}

\item{use.dip.lincomb}{DEFAULT: TRUE. Diploffect is parameterized based on additive haplotype effects
and heterozygous diplotype dominance effects. If posteriors on diplotypes are wanted, this option is required.
It does add to the computation time.}

\item{seed}{DEFAULT: 1. Diploffect involves a sampling process of the diplotypes, thus a seed is necessary
to produce the same results over multiple runs and different machines.}

\item{gamma.rate}{DEFAULT: 1. Hyper parameter for the variance component priors.}

\item{impute.on}{DEFAULT: "SUBJECT.NAME". Samples of diplotypes are drawn for each level of this variable.
Must be a column in the data frame. Allows the same diplotypes to be sampled if there are repeat observations
of a genome, such as in Collaborative Cross (CC) or CC-RIX.}

\item{weights.on}{DEFAULT: NULL. If numeric column in data frame is specified, weights are applied to the
diagonal of the random noise (diagonal) covariance matrix. This is an efficient approach to modeling
repeat observations, such as in CC or CC-RIX.}

\item{scale.response}{DEFAULT: TRUE. It is good to scale the response with INLA, thus the posteriors of effects
have a consistent scale (in standard deviations of the outcome).}

\item{do.augment.of.cache}{DEFAULT: FALSE. Not suggested. Could be used for CC genomecaches with only additive
dosages, though full probabilities are still greatly preferred.}
}
\value{
Diploffect.INLA object. Approximate posterior distributions for model parameters.
}
\description{
This function primarily takes a formula, locus, data frame, and genomecache and runs the
corresponding Diploffect model for a given number of samples.
}
\examples{
run.diploffect.inla.through.genomecache()
}
